tests/wpt/tests/css/css-ui/parsing/form-sizing-computed.html

<!DOCTYPE html>
<title>CSS UI: getComputedStyle().formSizing</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/pull/9251">
<link rel="author" title="Kent Tamura" href="mailto:tkent@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<body>
<div id="target"></div>
<script>
test_computed_value('form-sizing', 'auto');
test_computed_value('form-sizing', 'normal');
</script>
</body>
